Occupation System Change.
기존 1개인 줄 알았던 점령오브젝트를 여러개로 사용할 수 있게끔 수정.

+++ Source/Lakaya/Public/UI/TeamScoreWidget.h
@@ -36,8 +36,8 @@ public:
 	void ReMoveTeamScoreWidget(EOccupationGameState ChangeGameState);
 	
 	UFUNCTION()
-	void OnChangeATeamScore(uint8 NewScore);
+	void OnChangeATeamScore(float NewScore);
 	
 	UFUNCTION()
-	void OnChangeBTeamScore(uint8 NewScore);
+	void OnChangeBTeamScore(float NewScore);
 };
